+++ bpm-engine-api/src/main/java/io/takari/bpm/api/EventService.java
@@ -1,0 +1,13 @@
+package io.takari.bpm.api;
+
+import java.util.Collection;
+
+public interface EventService {
+
+    /**
+     * Returns a collection of currently waiting events for the specified process instance.
+     * @param processBusinessKey external business key of a process instance.
+     * @return collection of events.
+     */
+    Collection getEvents(String processBusinessKey);
+}